moved the existing graph algorithms into its own dedicated service in order to reuse and to customize them
